Gong  /Correct connection again/Calculate distance and relevant optimization

+++ app/src/main/java/net/kehui/www/t_907_origin/fragment/AdjustFragment.java
@@ -142,7 +142,7 @@ public class AdjustFragment extends Fragment {
                 }
                 break;
             case R.id.btn_vel_plus:
-                float velocity = ((MainActivity) getActivity()).getVelocity();
+                int velocity = ((MainActivity) getActivity()).getVelocity();
                 if (velocity < 250) {
                     velocity++;
                     ((MainActivity) getActivity()).setVelocity(velocity);
